[JS]Gather stamps tests. Extract verifying last stamp test to separated file.
